Improve Volley header handling.
-Add support for multiple headers with the same key. This is permitted
by HTTP standards - for example, a server may include multiple
Set-Cookie headers in a response. However, Volley used Map<String,
String> throughout its stack which made it impossible to see all but
the last such header. We now expose a flat Header list in
NetworkResponse and Cache.Entry for clients who wish to use it. The
old Map is still populated for backwards compatibility and
convenience.

-Fix inconsistent case-insensitivity in Header maps. Previously, the
header map would be case-insensitive for network requests but
case-sensitive for cache entry reads. We now ensure that the map is
always case-insensitive (for clients using it and not the flat list)
by creating it in NetworkResponse's constructor. Redundant TreeMap
creations have been removed.

-Avoid mutating the cache entry's response header map directly.

This change should be compile-time-compatible with existing Volley 1.0
users. There is a signature change to HttpResponse which has not been
part of an official release yet. Custom implementations of Cache may
wish to populate the new allHeaders field in Cache.Entry, though this
is optional.

Fixes #21
Fixes #76