Clarify/enforce immutability of header fields.
New fields in HttpResponse/NetworkResponse storing the full list of
headers are now documented and enforced as immutable (using
Collections#unmodifiableList). Note that this requires one additional
list copy in BasicNetwork for 304 handling.

The new header list field in Cache is documented as "should not be
mutated directly". While Volley's stock components can guarantee that
the list is immutable, we cannot guarantee that other caches don't
provide a mutable value.

The legacy Map<String, String> header fields are documented as
"should not be mutated directly" as well, as apps may be reliant on
mutating these fields directly given that they were previously
mutable (and were being mutated within Volley, even).