Workaround memory leak in dispatchers.
Android has a long-standing known issue where local variables aren't
explicitly cleared even when they go out of scope, which can cause
their contents to leak. Clearing the variable explicitly is not a
sufficient fix because the clear is optimized out as it's not
observable. The best we can do is extract the contents of the loops in
question to helper methods and hope that they are complex enough that
code optimizers like Proguard won't inline them.

Should contain no functional changes otherwise.

Verified against provided sample app.

Fixes #114