Workaround memory leak in dispatchers.
Android has a long-standing known issue where local variables aren't
explicitly cleared even when they go out of scope, which can cause
their contents to leak. Clearing the variable explicitly is not a
sufficient fix because the clear is optimized out as it's not
observable. Extracting the loop contents to a helper function is
likely to work but risks being inlined by a code optimizer. So we
poll() once (which is non-blocking) before waiting to ensure that the
previous request is always cleared expediently.

Should contain no functional changes otherwise.

Verified against provided sample app.

Fixes #114