Log fix (#112)
* Fix finding "caller" in VolleyLog

"trace[i].getClass()" always returns StackTraceElement, the class of
"trace[i]" itself. Method getClassName() is what really needed here to
get the class of element in the call stack trace.
Because "clazz" was alwasy StackTraceElement so the condition
"if (!clazz.equals(VolleyLog.class.getName()))" was always true. But
code worked with no problem because loop started from index 2
(int i = 2) which luckily is always the method called VolleyLog.

* Fix call to MarkerLog.finish() in Request

The "this" in run() referes to the anonymous Runnable containing it. The
intention here was to calling the toString() of the Request class.