Add a concept of public APIs to Volley.
-All public/protected classes/methods/fields are considered public
 API. VolleyApisTest uses reflection to find all public APIs and
 compare against a golden list so that APIs are not accidentally
 changed in the future.

-A new .internal package can be used for any classes which need to
 be public but should not be depended on by Volley clients. Any
 class in this package will be obfuscated by Proguard and excluded
 from Javadocs in release artifacts.

This is a relatively simple approach compared to the doclava-based
approach used by the support library / Android platform, which isn't
worth the complexity. However, one drawback is that this cannot hide
public/protected parts of classes in the core Volley package.
Unfortunately, Proguard does not appear to support such a directive
(short of manually listing everything that needs to be kept).

Fixes #102