Apply google-java-format to all code.
Also adds Gradle commands to format the code and verify formatting on
all files in the tree. CONTRIBUTING.md has been updated with the
steps to format + verify changes, and the format verification has
been added to the continuous integration setup (but if it fails, the
rest of the build will still be attempted to minimize churn).

Fixes #99