Cleaning up parameter comments + unnecessary generics.
Used diamond operator in all places it was possible.

For parameter comments, there's no way to automatically detect when
they "should" be present, but we can enforce that any that are there
adhere to the recommended style (which enables richer errorprone
checks). Took a manual pass to fix up some missing parameter tags,
but there may be others.

Diamond operators have been removed.

Test code is left unchanged to keep this tractable.

Fixes #166