Fix concurrent modification errors in ImageContainer. (#179)
ImageContainer#cancelRequest modifies data structures that are
mutated without a lock on the main thread. Its parent class,
ImageLoader, is documented as requiring that all access happen
on the main thread, so strictly enforce that cancelRequest be
called from the main thread as well. Outside of cancelRequest,
the only uncontrolled caller is NetworkImageView#setImageUrl,
which we also document/enforce as needing to be called from the
main thread.

Add @MainThread annotations and Javadoc clarifications as
appropriate. Clarify that behavior on custom ResponseDeliveries
is undefined.

Fixes #132