Re-raise swallowed interrupts and log warnings for spurious interrupts.
Volley relies on a quit() method and mQuit flag on both dispatcher
threads to shut down those threads. These could just use the interrupt
state instead, but as noted on #60, these are public, non-final
classes, and so relying on interrupt depends on external callers and
subclasses using the interrupt state correctly.

To be more conservative and retain existing behavior, we keep the quit
method/flag around to ensure that we only shut down the dispatchers
when quit() is called. However, we re-raise the interrupt flag in this
case. If we're interrupted outside of quit(), we suppress the
interrupt flag and log a warning instead.

Fixes #60