Don't drop batched responses when new requests complete.
If a request for a given URL is queued and completes while pending
responses are waiting for the next batch delivery window, all pending
responses are erroneously dropped on the floor because the pending
BatchedImageRequest is clobbered. Prevent this by checking for pending
batched responses when queueing new requests.

Fixes #209