Reduce cache churn from large entries.
If an entry is so large that adding it to the cache would trigger a
prune, and if that prune would clear the new entry as well, then skip
adding it in the first place.

For simplicity, we don't include the cache header overhead in this
calculation; this implies that we may still put and immediately clear
an entry if it is very close to the threshold. But this won't result
in incorrect behavior; it will just lead to a bit of file churn. We
could resolve this in the future by calculating the header overhead.